+++ app/src/main/java/com/linoop/bluefox/business/UserRepository.kt
@@ -2,6 +2,6 @@ package com.linoop.bluefox.business
 
 interface UserRepository {
     suspend fun createUser(userModel: UserModel):Long
-    suspend fun getUserById(userId: Int): UserModel
+    suspend fun getUserById(userId: Long): UserModel
     suspend fun getUsers(): List<UserModel>
 }
